HalfBasalTarget to Double (to support mmol/l)

+++ core/objects/src/main/kotlin/app/aaps/core/objects/extensions/ExtendedBolusExtension.kt
@@ -76,7 +76,7 @@ fun EB.iobCalc(
     profile: Profile,
     lastAutosensResult: AutosensResult,
     exerciseMode: Boolean,
-    halfBasalExerciseTarget: Int,
+    halfBasalExerciseTarget: Double,
     isTempTarget: Boolean,
     insulinInterface: Insulin
 ): IobTotal {
